requirements.txt
numpy
olefile
opencv-python
Pillow
qrcode
six
# zbarlight==1.2
pyzbar